Fix comparison functions
I had to remove the cast to bool operators, because in case of
comparison between pointers of non-matching types (e.g.
weak_ptr == owner_ptr) the compiler would choose bool == bool over
weak_ptr == weak_ptr. Unfortunately this forces the user to explicitly
compare with nullptr.

The std::rel_ops do not work when one of the operands needs to be
type-casted, so I had to provide manual implementations of != as well.